Refactor authentication and database session management to use environment variables for configuration
