+++ app/auth/utils/security.py
@@ -1,11 +1,15 @@
 # define functions for password hashing, token creation, and verfication 
+import os
 from datetime import datetime, timedelta
 from jose import jwt, JWTError
 from passlib.context import CryptContext
+from dotenv import load_dotenv
 
-SECRET_KEY = "your-secret-key"  # 実際は .env に置く
+load_dotenv()  # take environment variables from .env.
+
+SECRET_KEY = os.getenv("SECRET_KEY")
 ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 30
+ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30))
 
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
