+++ app/db/session.py
@@ -1,9 +1,17 @@
+import os
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
+from dotenv import load_dotenv
+from app.db.base import BaseModel
 
-DATABASE_URL = "postgresql+asyncpg://user:password@localhost:5432/dbname"
+load_dotenv()  # take environment variables from .env.
 
-engine = create_engine(DATABASE_URL, echo=True)  # 同期版なら `create_engine`
+DATABASE_URL = os.getenv("DATABASE_URL")
+engine = create_engine(DATABASE_URL, echo=True)
+
+BaseModel.metadata.create_all(bind=engine)
+
+
 SessionLocal = sessionmaker(bind=engine, autocommit=False, autoflush=False)
 
 def get_db():
